fix: removal of traits in identify events and lifecycle events (#8)
* fix: removal of traits in identify events

Removal of traits could have unwanted effect when there are multiple device mode integrations. Hence, it is safe not to remove them.

* fix: return after `Application Installed` or `Application Updated` event were made

Previously, as we are not returning (after making lifeCycle events), as a result, duplicate events are likely to be sent. To fix that, we need to return immediately.

* fix: reset log message

Added extra info for better visibility