feat!: update MoEngage SDK to the latest version 12.5.04 (#12)
* chore(sample-app): update version in build.gradle and gradle plugin version

* chore(moengage): bump to latest version of MoEngage sdk

* chore(sampple-app): add exported command in manifest file

* feat(moengage): update all the MoEngage APIs

Substitute MoEHelper (as it is deprecated) with MoEAnalyticsHelper class. Update RESET API (as per MoEngage doc). Add context parameter in every MoEngage API. Comment life cycle methods, as it is no longer supported.
API changed: `setEmailId`, `setUserName`, `setUserName`, `RESET`

* chore(moengage): remove comment (life cycle method) , un-imported class and addressed compiler warning

* fix(moengage): make `jsonToProperties` return non-null value

Compiler was throwing an error for trackEvent API, as it doesn't accept non-null value. So made it to return empty Properties() in case of any failure to convert json to Properties object.

* fix(sample-app): update several dependencies as per moengage requirement

Sample app was throwing error due to incorrect SdkVersion and plugin `com.google.gms.google-services` was not implemented and few other incorrect versions. Also added, MoEngage dependency in the sample app and kotlin in root build.gradle file.

* chore(sample-app): update code snippet in MainApplication.kt

Updated MoEngage SDK init code and made other minor improvement

* fix(moengage): update maven central publish scripts

* chore(moengage): bump version to `2.0.0`

* doc: update readme

* refactor: improve reset

* chore: change core android sdk version to `[1.7.0, 1.8.0)`

* fix: removal of traits in identify events and lifecycle events (#8)

* fix: removal of traits in identify events

Removal of traits could have unwanted effect when there are multiple device mode integrations. Hence, it is safe not to remove them.

* fix: return after `Application Installed` or `Application Updated` event were made

Previously, as we are not returning (after making lifeCycle events), as a result, duplicate events are likely to be sent. To fix that, we need to return immediately.

* fix: reset log message

Added extra info for better visibility

* fix: increase the moengage minSdkVersion to 21